pyproject.toml
[tool.poetry]
name = "tree-style-menu-django"
version = "0.1.0"
description = ""
authors = ["Deyvidas <devidbanyavichyus@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
django-debug-toolbar = "^4.3.0"


[tool.poetry.group.prod.dependencies]
django = "^5.0.4"
python-slugify = "^8.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
